Tier1: use signed size types for buffer size field

+++ src/public/tier1/cmd.h
@@ -73,7 +73,7 @@ public:
 private:
 	cmd_source_t m_nQueuedVal;
 	int          m_nArgc;
-	int64_t      m_nArgv0Size;
+	ssize_t      m_nArgv0Size;
 	char         m_pArgSBuffer[COMMAND_MAX_LENGTH];
 	char         m_pArgvBuffer[COMMAND_MAX_LENGTH];
 	const char*  m_ppArgv[COMMAND_MAX_ARGC];
